pyproject.toml
[tool.poetry]
name = "oshinglish"
version = "0.1.0"
description = "A Django-built web application for an online English-Oshindonga dictionary"
authors = ["abiatarfestus <40578707+abiatarfestus@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.9.1"
Django = "^4.2.5"
djangorestframework = "^3.14.0"
isort = "^5.12.0"
mysqlclient = "^2.2.0"
django-debug-toolbar = "^4.2.0"
python-decouple = "^3.8"
Pillow = "^10.0.1"
whitenoise = "^6.5.0"
django-simple-history = "^3.4.0"
django-resized = "^1.0.2"
six = "^1.16.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
